Added BlackHoleTests.
