+++ BlackHole/BlackHole.plist
@@ -21,7 +21,7 @@
 	<key>CFBundleSignature</key>
 	<string>????</string>
 	<key>CFBundleVersion</key>
-	<string>531</string>
+	<string>547</string>
 	<key>CFPlugInFactories</key>
 	<dict>
 		<key>e395c745-4eea-4d94-bb92-46224221047c</key>
